refactor app.py (1/7): extract auth helpers to helpers/auth.py
is_authenticated() and cd_auth_required() decorator move from app.py
to a new helpers/ package. app.py imports them back. New helpers/
package created with empty __init__.py.

Smoke-tested: /login renders, POST /login authenticates and redirects,
authed /publish + /below-deck + /command-deck/ (the last exercising
cd_auth_required) all return 200.

First sub-phase of 7 in the helpers extraction. Pattern proven —
remaining helpers follow the same cadence: move function(s) to
helpers/<name>.py, replace defs in app.py with `from helpers.<name>
import …`, smoke-test routes that exercise the moved code.
